+++ src/completions/suggestion_utils.cpp
@@ -232,16 +232,16 @@ int edit_distance(const std::string& str1, const std::string& str2) {
     const size_t n = str2.length();
 
     if (m == 0)
-        return n;
+        return static_cast<int>(n);
     if (n == 0)
-        return m;
+        return static_cast<int>(m);
 
     std::vector<std::vector<int>> dp(m + 1, std::vector<int>(n + 1));
 
     for (size_t i = 0; i <= m; i++)
-        dp[i][0] = i;
+        dp[i][0] = static_cast<int>(i);
     for (size_t j = 0; j <= n; j++)
-        dp[0][j] = j;
+        dp[0][j] = static_cast<int>(j);
 
     for (size_t i = 1; i <= m; i++) {
         for (size_t j = 1; j <= n; j++) {
